+++ CHANGELOG.md
@@ -7,6 +7,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.61.0] - 2021-08-15
+
+- fixed command help display (incorrect key, was hard coded to `gunner`)
+- Upgraded @codedungeon/messenger to v0.14.0
+- Fixed issue with `make-command.mustache` template, should include require of `@codedungeon/gunner`
+- Fixed issue when running cli for first time after install (create commands directory if not exists)
+
 ## [0.57.0 .. 0.57.1] - 2021-07-23
 
 - Fixed assumption that `commandName` would use `input` prompt